refactor: move css files into separate modules to make css changes simpler
