+++ packages/renderer-worker/src/parts/ViewletActivityBar/ViewletActivityBarCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletActivityBar.css', '/css/parts/MaskIcon.css']
