+++ packages/renderer-worker/src/parts/ViewletAudio/ViewletAudioCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletAudio.css']
