+++ packages/renderer-worker/src/parts/ViewletColorPicker/ViewletColorPickerCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletColorPicker.css']
