+++ packages/renderer-worker/src/parts/ViewletDebugConsole/ViewletDebugConsole.ipc.js
@@ -2,7 +2,6 @@ export const name = 'Debug Console'
 
 export const Commands = {}
 
-export const Css = '/css/parts/ViewletDebugConsole.css'
-
+export * from './ViewletDebugConsoleCss.js'
 export * from './ViewletDebugConsole.js'
 export * from './ViewletDebugConsoleActions.js'
