+++ packages/renderer-worker/src/parts/ViewletDebugConsole/ViewletDebugConsoleCss.js
@@ -1,0 +1,1 @@
+export const Css = '/css/parts/ViewletDebugConsole.css'
