+++ packages/renderer-worker/src/parts/ViewletDialog/ViewletDialogCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletDialog.css', '/css/parts/IconButton.css']
