+++ packages/renderer-worker/src/parts/ViewletDiffEditor/ViewletDiffEditorCss.js
@@ -1,0 +1,4 @@
+export const Css = [
+  '/css/parts/ViewletDiffEditor.css',
+  '/css/parts/ViewletEditor.css',
+]
