+++ packages/renderer-worker/src/parts/ViewletEditorCompletion/ViewletEditorCompletion.ipc.js
@@ -15,6 +15,5 @@ export const LazyCommands = {
   ...VirtualList.LazyCommands
 }
 
-export const Css = '/css/parts/ViewletEditorCompletion.css'
-
+export * from './ViewletEditorCompletionCss.js'
 export * from './ViewletEditorCompletion.js'
