+++ packages/renderer-worker/src/parts/ViewletEditorCompletion/ViewletEditorCompletionCss.js
@@ -1,0 +1,1 @@
+export const Css = '/css/parts/ViewletEditorCompletion.css'
