+++ packages/renderer-worker/src/parts/ViewletEditorImage/ViewletEditorImageCss.js
@@ -1,0 +1,1 @@
+export const Css = '/css/parts/ViewletEditorImage.css'
