+++ packages/renderer-worker/src/parts/ViewletEditorText/ViewletEditorTextCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletEditor.css', '/css/parts/ScrollBar.css']
