+++ packages/renderer-worker/src/parts/ViewletExplorer/ViewletExplorerCss.js
@@ -1,0 +1,7 @@
+export const Css = [
+  '/css/parts/ViewletExplorer.css',
+  '/css/parts/Label.css',
+  '/css/parts/TreeItem.css',
+  '/css/parts/Drag.css',
+  '/css/parts/InputBox.css',
+]
