+++ packages/renderer-worker/src/parts/ViewletExtensionDetail/ViewletExtensionDetailCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletExtensionDetail.css', '/css/parts/Markdown.css']
