+++ packages/renderer-worker/src/parts/ViewletExtensions/ViewletExtensionsCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletExtensions.css', '/css/parts/ViewletList.css', '/css/parts/InputBox.css', '/css/parts/ScrollBar.css']
