+++ packages/renderer-worker/src/parts/ViewletFindWidget/ViewletFindWidget.ipc.js
@@ -9,11 +9,5 @@ export const Commands = {
   handleInput: ViewletFindWidget.handleInput,
 }
 
-export const Css = [
-  '/css/parts/ViewletFindWidget.css',
-  '/css/parts/IconButton.css',
-  '/css/parts/MaskIcon.css',
-  '/css/parts/InputBox.css',
-]
-
+export * from './ViewletFindWidgetCss.js'
 export * from './ViewletFindWidget.js'
