+++ packages/renderer-worker/src/parts/ViewletKeyBindings/ViewletKeyBindingsCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletKeyBindings.css', '/css/parts/InputBox.css']
