+++ packages/renderer-worker/src/parts/ViewletOutput/ViewletOutputCss.js
@@ -1,0 +1,1 @@
+export const Css = '/css/parts/ViewletOutput.css'
