+++ packages/renderer-worker/src/parts/ViewletPanel/ViewletPanelCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletPanel.css', '/css/parts/IconButton.css', '/css/parts/Actions.css']
