+++ packages/renderer-worker/src/parts/ViewletSearch/ViewletSearchCss.js
@@ -1,0 +1,9 @@
+export const Css = [
+  '/css/parts/Highlight.css',
+  '/css/parts/InputBox.css',
+  '/css/parts/Label.css',
+  '/css/parts/ViewletList.css',
+  '/css/parts/ViewletSearch.css',
+  '/css/parts/TreeItem.css',
+  '/css/parts/ScrollBar.css',
+]
