+++ packages/renderer-worker/src/parts/ViewletSideBar/ViewletSideBar.ipc.js
@@ -9,6 +9,5 @@ export const Commands =  {
   show: ViewletSideBar.openViewlet,
 }
 
-export const Css = ['/css/parts/ViewletSideBar.css', '/css/parts/Actions.css', '/css/parts/IconButton.css']
-
+export * from './ViewletSideBarCss.js'
 export * from './ViewletSideBar.js'
