+++ packages/renderer-worker/src/parts/ViewletSideBar/ViewletSideBarCss.js
@@ -1,0 +1,1 @@
+export const Css = ['/css/parts/ViewletSideBar.css', '/css/parts/Actions.css', '/css/parts/IconButton.css']
