+++ packages/renderer-worker/src/parts/ViewletTitleBarMenuBar/ViewletTitleBarMenuBarCss.js
@@ -1,0 +1,1 @@
+export const Css = '/css/parts/ViewletTitleBarMenuBar.css'
